Improves Logic for Finding Theme Files
When trying to build the theme locally, I noticed that many files were not included in the original package. That was the case because the wild cards used in `setup.py` did not include sub-directories in the `static` path. That resulted in fonts and vendor JavaScript not being included in the final build, breaking the theme.

I solved that by adding a new recursive function that adds all paths from a given base path. In this case, it will distribute the theme package with all the files under the `static` directory.